customer crud

+++ src/main/java/com/example/springapi/model/repository/CustomerRepository.java
@@ -3,5 +3,5 @@ package com.example.springapi.model.repository;
 import com.example.springapi.model.entity.Customers;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomerRepository extends JpaRepository<Customers,Integer> {
+public interface CustomerRepository extends JpaRepository<Customers,Long> {
 }
