+++ src/main/java/com/example/springapi/model/repository/IngredientsRepository.java
@@ -3,5 +3,5 @@ package com.example.springapi.model.repository;
 import com.example.springapi.model.entity.Ingredients;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IngredientsRepository extends JpaRepository<Ingredients,Integer> {
+public interface IngredientsRepository extends JpaRepository<Ingredients,Long> {
 }
