+++ src/main/java/com/example/springapi/model/repository/StockRepository.java
@@ -3,5 +3,5 @@ package com.example.springapi.model.repository;
 import com.example.springapi.model.entity.Stocks;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StockRepository extends JpaRepository<Stocks,Integer> {
+public interface StockRepository extends JpaRepository<Stocks,Long> {
 }
